# swagger-spec.json
{"openapi":"3.0.0","paths":{"/event":{"get":{"operationId":"getEvents","summary":"returns all events in database","tags":["event"],"parameters":[],"responses":{"200":{"description":"array of events","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDTOResponse"}}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}},"post":{"operationId":"createEvent","summary":"create event contract and add it to the database","tags":["event"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDTORequest"}}}},"responses":{"200":{"description":"return id of created event","content":{"application/json":{"schema":{"type":"number"}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/event/events_by_tg/{id}":{"get":{"operationId":"getEventsById","summary":"get events by user telegram id","tags":["event"],"parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"return array of events","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDTOResponse"}}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/event/events_by_name/{name}":{"get":{"operationId":"getEventsByName","summary":"get events by name","tags":["event"],"parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"return array of events","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventDTOResponse"}}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/event/event_by_id/{id}":{"get":{"operationId":"getEventByID","summary":"get event by id","tags":["event"],"parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"return an event or null","content":{"application/json":{"schema":{"$ref":"#/components/schemas/EventDTOResponse"}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/event/approveLink":{"post":{"operationId":"updateApproveLink","summary":"update approve link by id","tags":["event"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"example":{"event_id":2,"approve_link":"link"}}}}},"responses":{"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/person":{"post":{"operationId":"createPerson","summary":"create person","tags":["person"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}}},"responses":{"200":{"description":"return id of created person","content":{"application/json":{"schema":{"type":"number"}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/person/person_by_tg/{id}":{"get":{"operationId":"getPersonById","summary":"get person and his roles by tg if no role return null","tags":["person"],"parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":[{"id":1,"walletAddress":"address","roles":[{"name":"ADMIN"},{"name":"USER"}]}]}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/participantList/register":{"post":{"operationId":"registerPersonOnEvent","summary":"register user on event","tags":["participantList"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"example":{"person_id":1,"event_id":2}}}}},"responses":{"200":{"description":"return true if created in another case false","content":{"application/json":{"schema":{"type":"boolean"}}}}}}},"/participantList/approve":{"post":{"operationId":"approveParticipation","summary":"approve participant of user","tags":["participantList"],"parameters":[{"name":"person_id","required":true,"in":"path","schema":{"type":"number"}},{"name":"event_id","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"example":{"person_id":1,"event_id":2}}}}},"responses":{"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/participantList/mint/{event_id}":{"post":{"operationId":"mint","summary":"mint nft after event","tags":["participantList"],"parameters":[{"name":"event_id","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/location":{"get":{"operationId":"getLocations","summary":"get locations","tags":["location"],"parameters":[],"responses":{"200":{"description":"return array of locations","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocationDTO"}}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}},"post":{"operationId":"createLocation","summary":"create location","tags":["location"],"parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDTO"}}}},"responses":{"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/type":{"get":{"operationId":"getTypes","summary":"get types","tags":["type"],"parameters":[],"responses":{"200":{"description":"return array of types","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventTypeDTO"}}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}},"/status":{"get":{"operationId":"getStatuses","summary":"get statuses of event","tags":["status"],"parameters":[],"responses":{"200":{"description":"return array of statuses","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/EventStatusDTO"}}}}},"400":{"description":"returns error message if  PrismaClientKnownRequestError, PrismaClientUnknownRequestError, PrismaClientValidationError","content":{"application/json":{"schema":{"example":{"message":"message"}}}}},"500":{"description":"returns error message if PrismaClientRustPanicError, PrismaClientInitializationError, or smth unknown else","content":{"application/json":{"schema":{"example":{"message":"message"}}}}}}}}},"info":{"title":"NFT-Tickets server","description":"Main server","version":"0.1","contact":{}},"tags":[],"servers":[{"url":"https://prisma-production-1488.up.railway.app"}],"components":{"schemas":{"EventDTOResponse":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"isSBT":{"type":"boolean"},"nftIpfsUrl":{"type":"string","uniqueItems":true},"collectionAddr":{"type":"string"},"urlCover":{"type":"string"},"description":{"type":"string"},"creatorID":{"type":"number"},"started_at":{"format":"date-time","type":"string"},"finished_at":{"format":"date-time","type":"string"},"locationID":{"type":"number"},"registeredParticipants":{"type":"number"},"countOfRewardTokens":{"type":"number"},"typeId":{"type":"number"},"statusId":{"type":"number"}},"required":["id","name","isSBT","nftIpfsUrl","collectionAddr","creatorID","registeredParticipants","countOfRewardTokens","typeId","statusId"]},"EventDTORequest":{"type":"object","properties":{"name":{"type":"string"},"urlCover":{"type":"string","description":"ask Yaroslav, don't care now"},"description":{"type":"string"},"creatorTgId":{"type":"string","description":"telegram id of creator"},"isSBT":{"type":"boolean"},"started_at":{"format":"date-time","type":"string"},"finished_at":{"format":"date-time","type":"string"},"locationID":{"type":"number"},"nftIpfsUrl":{"type":"string","uniqueItems":true,"description":"nft ipfs url"},"countOfRewardTokens":{"type":"number"},"typeId":{"type":"number","default":1,"description":"type of event, default not stated"},"statusId":{"type":"number","default":1,"description":"status of event, default not registered"}},"required":["name","creatorTgId","isSBT","nftIpfsUrl","countOfRewardTokens"]},"PersonDTO":{"type":"object","properties":{"walletAddress":{"type":"string"},"tgId":{"type":"string","description":"telegram id"}},"required":["walletAddress","tgId"]},"LocationDTO":{"type":"object","properties":{"address":{"type":"string"},"room":{"type":"string"},"isOffline":{"type":"boolean"},"link":{"type":"string"}},"required":["address","room","isOffline","link"]},"EventTypeDTO":{"type":"object","properties":{"type":{"type":"string"}},"required":["type"]},"EventStatusDTO":{"type":"object","properties":{"status":{"type":"string"}},"required":["status"]}}}}